requirements.txt
# requirements.txt - Совместимые версии
telethon==1.34.0
python-telegram-bot==20.7
httpx==0.25.2
aiohttp==3.9.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
pillow==10.3.0